+++ tests/prompt/bp/prompt-bp-uncheckmark_test.py
@@ -486,7 +486,7 @@ class TestDynamicNodes:
     def test_checkmark_by_obj1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
 
-        node = bp.corpus["(Abbreviations)"]
+        node = bp.corpus["(Decode-Only Shorthand)"]
 
         assert bp.uncheckmark(node)
         assert not bp.is_checkmarked(node)
@@ -494,7 +494,7 @@ class TestDynamicNodes:
     def test_checkmark_by_hash1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
 
-        node = bp.corpus["(Abbreviations)"]
+        node = bp.corpus["(Decode-Only Shorthand)"]
         node_hash = hash(node)
 
         assert bp.uncheckmark(node_hash)
@@ -503,8 +503,8 @@ class TestDynamicNodes:
 
     def test_checkmark_by_name1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
-        node = bp.corpus["(Abbreviations)"]
-        ipt = "(Abbreviations)"
+        node = bp.corpus["(Decode-Only Shorthand)"]
+        ipt = "(Decode-Only Shorthand)"
 
         assert bp.uncheckmark(ipt)
 
@@ -512,8 +512,8 @@ class TestDynamicNodes:
 
     def test_checkmark_by_identifier1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
-        node = bp.corpus["(Abbreviations)"]
-        ipt = "(Abbreviations)"
+        node = bp.corpus["(Decode-Only Shorthand)"]
+        ipt = "(Decode-Only Shorthand)"
 
         assert bp.uncheckmark(ipt)
 
@@ -574,7 +574,7 @@ class TestDynamicNodes:
 [ ] │   │           └── Future Work
 [ ] │   └── Conclusion
 [x] ├── (Today)
-[x] ├── (Abbreviations)
+[x] ├── (Decode-Only Shorthand)
 [x] ├── (usable-abbreviations)
 [x] ├── (natural-language-codes)
 [x] └── (programming-language-codes)"""
@@ -601,7 +601,7 @@ class TestDynamicNodes:
 [x] │   │           └── Future Work
 [x] │   └── Conclusion
 [x] ├── (Today)
-[x] ├── (Abbreviations)
+[x] ├── (Decode-Only Shorthand)
 [x] ├── (usable-abbreviations)
 [x] ├── (natural-language-codes)
 [x] └── (programming-language-codes)"""